pyproject.toml
[tool.poetry]
name = "solafune-mine-detection"
version = "0.1.0"
description = "This repository is dedicated to the Mine Detection competition on Solafune."
authors = ["mfujiwara"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.0"
scikit-learn = "^1.4.0"
mlflow = "^2.10.1"
hydra-core = "^1.3.2"
matplotlib = "^3.8.2"
seaborn = "^0.13.2"
opencv-python-headless = "^4.9.0.80"
albumentations = "^1.3.1"
timm = "^0.9.12"
torch = "2.1.2"
rasterio = "^1.3.9"
tqdm = "^4.66.1"
ipywidgets = "^8.1.1"
transformers = "^4.37.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
